Moved and improved test.
