+++ community/cypher/acceptance-spec-suite/src/test/scala/org/neo4j/internal/cypher/acceptance/IndexNestedLoopJoinAcceptanceTest.scala
@@ -74,22 +74,6 @@ class IndexNestedLoopJoinAcceptanceTest extends ExecutionEngineFunSuite with New
     result should not(use("ValueHashJoin", "CartesianProduct", "NodeByLabelScan", "Filter"))
   }
 
-  test("should use index on literal map expression") {
-    val nodes = Range(0,125).map(i => createLabeledNode(Map("id" -> i), "Foo"))
-    graph.createIndex("Foo", "id")
-    val query =
-      """
-        |PROFILE
-        | MATCH (f:Foo)
-        | USING INDEX f:Foo(id)
-        | WHERE f.id={id: 123}.id
-        | RETURN f
-      """.stripMargin
-    val result = executeWithCostPlannerOnly(query)
-    result.columnAs[Node]("f").toSet should equal(Set(nodes(123)))
-    result.executionPlanDescription() should includeAtLeastOne(classOf[NodeIndexSeek], withVariable = "f")
-  }
-
   test("should use index on variable defined from literal map") {
     val nodes = Range(0,125).map(i => createLabeledNode(Map("id" -> i), "Foo"))
     graph.createIndex("Foo", "id")
